+++ Semester_2/Lab_9/polynom.h
@@ -1,11 +1,11 @@
 struct polynom;
 
-polynom *new_polynom(int *arr, int n);
-void upd_polynom(polynom *p, int *arr, int n);
-polynom *new_binom(int a);
-void upd_binom(polynom *b, int a);
+polynom *new_polynom(long long *arr, unsigned int n);
+void upd_polynom(polynom *p, long long *arr, unsigned int n);
+polynom *new_binom(long long a);
+void upd_binom(polynom *b, long long a);
 polynom *multiply(polynom *a, polynom *b);
 void multiply(polynom *m, polynom *a, polynom *b);
 void print_polynom(polynom *p);
-void print_koeffs(int *arr, int n);
+void print_koeffs(long long *arr, unsigned int n);
 void destroy(polynom *p);
